SLIM-2253-use-secret-mangement-for-secure-keys
