+++ osgp/platform/osgp-secret-management/src/main/resources/application.properties
@@ -1,7 +1,7 @@
 #Security configuration
 component.name=osgp-secret-management
-database.secret.resource=file:/etc/ssl/certs/secret.aes
-soap.secret.resource=file:/etc/ssl/certs/soap-secret.aes
+database.secret.resource=file:/etc/osp/smartmetering/keys/platform/secret-mgmt-db.key
+soap.secret.resource=file:/etc/osp/smartmetering/keys/platform/secret-mgmt-soap.key
 encryption.provider.type=HSM
 
 #hsm.keystore.resource=file:/etc/ssl/certs/houston.keystore
