--- osgp/protocol-adapter-dlms/osgp-protocol-adapter-dlms/src/main/java/org/opensmartgridplatform/adapter/protocol/dlms/application/services/ISecurityKeyService.java
@@ -1,8 +1,0 @@
-package org.opensmartgridplatform.adapter.protocol.dlms.application.services;
-
-public interface ISecurityKeyService {
-
-    byte[] getDlmsAuthenticationKey(final String deviceIdentification);
-    byte[] getDlmsGlobalUnicastEncryptionKey(final String deviceIdentification);
-
-}
