--- osgp/protocol-adapter-dlms/osgp-protocol-adapter-dlms/src/main/java/org/opensmartgridplatform/adapter/protocol/dlms/application/services/SecretMangementKeyService.java
@@ -1,14 +1,0 @@
-package org.opensmartgridplatform.adapter.protocol.dlms.application.services;
-
-public class SecretMangementKeyService implements ISecurityKeyService {
-
-    @Override
-    public byte[] getDlmsAuthenticationKey(String deviceIdentification) {
-        return new byte[0];
-    }
-
-    @Override
-    public byte[] getDlmsGlobalUnicastEncryptionKey(String deviceIdentification) {
-        return new byte[0];
-    }
-}
