+++ osgp/protocol-adapter-dlms/osgp-protocol-adapter-dlms/src/main/resources/wsdl/SecretManagement.wsdl
@@ -1,0 +1,76 @@
+<?xml version="1.0" encoding="UTF-8" standalone="no"?>
+<wsdl:definitions xmlns:wsdl="http://schemas.xmlsoap.org/wsdl/" xmlns:sch="http://www.opensmartgridplatform.org/schemas/security/secretmanagement/2020/05" xmlns:soap="http://schemas.xmlsoap.org/wsdl/soap/" xmlns:tns="http://www.opensmartgridplatform.org/schemas/security/secretmanagement/2020/05" targetNamespace="http://www.opensmartgridplatform.org/schemas/security/secretmanagement/2020/05">
+    <wsdl:types>
+        <xsd:schema xmlns:xsd="http://www.w3.org/2001/XMLSchema" attributeFormDefault="unqualified" elementFormDefault="qualified" targetNamespace="http://www.opensmartgridplatform.org/schemas/security/secretmanagement/2020/05"><xsd:simpleType name="Identification"><xsd:restriction base="xsd:normalizedString"><xsd:minLength value="1"/><xsd:maxLength value="40"/><xsd:pattern value="[^ ]{0,40}"/></xsd:restriction></xsd:simpleType><xsd:simpleType name="DeviceIdentification"><xsd:restriction base="tns:AlphaNumericType"><xsd:length value="17"/></xsd:restriction></xsd:simpleType><xsd:simpleType name="AlphaNumericType"><xsd:restriction base="xsd:string"><xsd:pattern value="[a-zA-Z0-9 ]+"/></xsd:restriction></xsd:simpleType><xsd:simpleType name="OsgpResultType"><xsd:restriction base="xsd:string"><xsd:enumeration value="OK"/><xsd:enumeration value="NOT FOUND"/><xsd:enumeration value="NOT OK"/></xsd:restriction></xsd:simpleType><xsd:element name="TechnicalFault"><xsd:complexType><xsd:sequence><xsd:element name="Message" type="xsd:string"/><xsd:element name="Component" type="xsd:string"/><xsd:element name="InnerException" type="xsd:string"/><xsd:element name="InnerMessage" type="xsd:string"/></xsd:sequence></xsd:complexType></xsd:element><xsd:simpleType name="SecretType"><xsd:restriction base="xsd:string"><xsd:enumeration value="E_METER_MASTER_KEY"/><xsd:enumeration value="E_METER_AUTHENTICATION_KEY"/><xsd:enumeration value="E_METER_ENCRYPTION_KEY_UNICAST"/><xsd:enumeration value="E_METER_ENCRYPTION_KEY_BROADCAST"/><xsd:enumeration value="G_METER_MASTER_KEY"/><xsd:enumeration value="G_METER_ENCRYPTION_KEY"/><xsd:enumeration value="G_METER_FIRMWARE_UPDATE_AUTHENTICATION_KEY"/><xsd:enumeration value="G_METER_OPTICAL_PORT_KEY"/><xsd:enumeration value="PPP_PASSWORD"/></xsd:restriction></xsd:simpleType><xsd:complexType name="SecretTypes"><xsd:sequence><xsd:element maxOccurs="16" name="SecretType" type="tns:SecretType"/></xsd:sequence></xsd:complexType><xsd:complexType name="TypedSecret"><xsd:sequence><xsd:element name="Type" type="tns:SecretType"/><xsd:element name="Secret" type="xsd:string"/></xsd:sequence></xsd:complexType><xsd:complexType name="TypedSecrets"><xsd:sequence maxOccurs="16"><xsd:element name="TypedSecret" type="tns:TypedSecret"/></xsd:sequence></xsd:complexType><xsd:element name="getSecretsRequest"><xsd:complexType><xsd:sequence><xsd:element name="DeviceId" type="tns:DeviceIdentification"/><xsd:element name="SecretTypes" type="tns:SecretTypes"/></xsd:sequence></xsd:complexType></xsd:element><xsd:element name="getSecretsResponse"><xsd:complexType><xsd:sequence><xsd:element name="Result" type="tns:OsgpResultType"/><xsd:element name="TypedSecrets" type="tns:TypedSecrets"/><xsd:element name="TechnicalFault"/></xsd:sequence></xsd:complexType></xsd:element><xsd:element name="storeSecretsRequest"><xsd:complexType><xsd:sequence><xsd:element name="DeviceId" type="tns:DeviceIdentification"/><xsd:sequence><xsd:element name="TypedSecrets" type="tns:TypedSecrets"/></xsd:sequence></xsd:sequence></xsd:complexType></xsd:element><xsd:element name="storeSecretsResponse"><xsd:complexType><xsd:sequence><xsd:element name="Result" type="tns:OsgpResultType"/><xsd:element name="TechnicalFault"/></xsd:sequence></xsd:complexType></xsd:element></xsd:schema>
+    </wsdl:types>
+    <wsdl:message name="getSecretsResponse">
+        <wsdl:part element="tns:getSecretsResponse" name="getSecretsResponse">
+        </wsdl:part>
+    </wsdl:message>
+    <wsdl:message name="storeSecretsResponse">
+        <wsdl:part element="tns:storeSecretsResponse" name="storeSecretsResponse">
+        </wsdl:part>
+    </wsdl:message>
+    <wsdl:message name="TechnicalFault">
+        <wsdl:part element="tns:TechnicalFault" name="TechnicalFault">
+        </wsdl:part>
+    </wsdl:message>
+    <wsdl:message name="getSecretsRequest">
+        <wsdl:part element="tns:getSecretsRequest" name="getSecretsRequest">
+        </wsdl:part>
+    </wsdl:message>
+    <wsdl:message name="storeSecretsRequest">
+        <wsdl:part element="tns:storeSecretsRequest" name="storeSecretsRequest">
+        </wsdl:part>
+    </wsdl:message>
+    <wsdl:portType name="SecretManagementPort">
+        <wsdl:operation name="getSecrets">
+            <wsdl:input message="tns:getSecretsRequest" name="getSecretsRequest">
+            </wsdl:input>
+            <wsdl:output message="tns:getSecretsResponse" name="getSecretsResponse">
+            </wsdl:output>
+        </wsdl:operation>
+        <wsdl:operation name="storeSecrets">
+            <wsdl:input message="tns:storeSecretsRequest" name="storeSecretsRequest">
+            </wsdl:input>
+            <wsdl:output message="tns:storeSecretsResponse" name="storeSecretsResponse">
+            </wsdl:output>
+        </wsdl:operation>
+        <wsdl:operation name="Technical">
+            <wsdl:fault message="tns:TechnicalFault" name="TechnicalFault">
+            </wsdl:fault>
+        </wsdl:operation>
+    </wsdl:portType>
+    <wsdl:binding name="SecretManagementPortSoap11" type="tns:SecretManagementPort">
+        <soap:binding style="document" transport="http://schemas.xmlsoap.org/soap/http"/>
+        <wsdl:operation name="getSecrets">
+            <soap:operation soapAction=""/>
+            <wsdl:input name="getSecretsRequest">
+                <soap:body use="literal"/>
+            </wsdl:input>
+            <wsdl:output name="getSecretsResponse">
+                <soap:body use="literal"/>
+            </wsdl:output>
+        </wsdl:operation>
+        <wsdl:operation name="storeSecrets">
+            <soap:operation soapAction=""/>
+            <wsdl:input name="storeSecretsRequest">
+                <soap:body use="literal"/>
+            </wsdl:input>
+            <wsdl:output name="storeSecretsResponse">
+                <soap:body use="literal"/>
+            </wsdl:output>
+        </wsdl:operation>
+        <wsdl:operation name="Technical">
+            <soap:operation soapAction=""/>
+            <wsdl:fault name="TechnicalFault">
+                <soap:fault name="TechnicalFault" use="literal"/>
+            </wsdl:fault>
+        </wsdl:operation>
+    </wsdl:binding>
+    <wsdl:service name="SecretManagementPortService">
+        <wsdl:port binding="tns:SecretManagementPortSoap11" name="SecretManagementPortSoap11">
+            <soap:address location="http://localhost:8080/osgp-secret-management/ws/SecretManagement"/>
+        </wsdl:port>
+    </wsdl:service>
+</wsdl:definitions>
